berkeley-abc_git.bb: Update to newest revision
Switch to using the cmake build, as it resolved link object dependence
correctly (or at least does not set incorrect link object dependence).

Switching to cmake requires minor change to how the ARCHFLAGS are passed
in. This is due to the cmake config wrapping the standard makefile by
calling cmake_info target, but does not allow the passing of extra args
to that make call.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>