recipes-pnr/nextpnr.bb: Fix up handling of bbasm on cross targets
The build of nextpnr does some processing on the host with the data from
the icestorm project. In order for this to work it needs to be able to
use it bbasm tool, but the build does not handle this automatically and
thus depends on the nextpnr-native build for the tool.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>