recipes-pnr/arachne-pnr_git.bb: Hide compiler details from version string
The version string built into the tool is used to match against the
chipdb built binaries. However in order to support cross builds the
tool must execute on the host which is likely to have a different
compiler to the target as well as different compiler flags. Remove the
compiler information from the version string to work around the
limitation.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>