symbiflow-arch-defs: Add definitions and tests recipes
Currently two separate recipes are added to handle generating
definitions and to run the symbiflow-arch-defs various tests.

Not all tests currently run, specifically the build only runs "CI"
tests, of which some currently fail. However the definitions are
generated and installed.

Also bump the version of verilog-to-routing to the "master+wip" head
revision.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>