WIP: fetch-restricted.bbclass: Allow a recipe to mark uris as restricted
Marking a URI as restricted will prevent the fetcher from attempting to
fetch the remote file. It will however allow the fetcher to get the file
from a PREMIRROR or handle the file as fetched if already located in the
'downloads/' directory.

If the fetch fails it will display an error to the user and point them
at the URI in which to download the file that has some restriction on
being able to download it.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
---
TODO: move this to meta-hdl-vendor?