WIP: vivado: Create two recipes to handle Xilinx Vivado tools
The 'vivado-native' recipe will install the tools into the sysroot and
provide them within the sysroot. This allows for automated install and
execution. There are some draw backs to this method, mainly that there
is duplication of the install within the sysroot and in sstate cache.

The second recipe 'vivado-external-native' relies on the install to be
provided from outside the OE build environment. This is useful on
development/local machines where the tools are already installed. The
sysroot is populated with wrapper shell scripts that allow for recipes
to use the same sysroot path with the scripts handling the execution of
the tools elsewhere in the system.

When depending on these tools a recipe should depend on the virtual
provider like so:

  DEPENDS += "virtual/xilinx-vivado-native"

The recipe should either execute the vivado tool from the PATH or can
provide the '${RECIPE_SYSROOT_NATIVE}/opt/xilinx/*' path into tools to
reference the Vivado install path.

By default to any of these recipes a user must agree and whitelist (in
LICENSE_FLAGS_WHITELIST) the following licenses:

  xilinx
  xilinx-webpack
  xilinx-3rd-party

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
---
TODO: move this into a meta-hdl-vendor layer?