tinyfpga-programmer: Add recipe and dependencies
Add a recipe for the CLI programmer part of the TinyFPGA bootloader
loader project.

In order to use this recipe a number of other python dependencies needed
to be added as recipes as well as updated versions of existing recipes
(e.g. setuptools-scm) and appends to enable native and nativesdk
extended classes for pyserial and pyusb. These recipes have be added
into the same directory as the tinyfpga recipe in order to manage them
within the meta-hdl layer.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>