litex-linux-vexriscv: Add recipes for litex-linux-vexriscv
Add recipe for litex-linux-vexriscv, and a machine configuration for
the versa-ecp5 target. Additionally add a recipe for the kernel
configured for this target only based on a mainline kernel at 5.4-rc7.

This layer does not include the all the parts needed to build linux
userspace components for this target, the meta-riscv layer is required
for the associated 32-bit RISC-V support.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>