linux-on-litex-vexriscv.bb: Fix clean build from sstate issues
Change create_toolchain_wrappers to happen before do_configure. This
ensures that the wrappers are available for the configure task, but also
makes it easier to depend on the wrappers without explicitly depending
on the task.

Fix task dependence on core-image-minimal, do_image_cpio is too early to
have the deployed artifacts available in the images output. Instead
depend on do_image_complete, this comes with the drawback of having the
EXTRA_IMAGEDEPENDS tasks being executed (aka bitstream compilation).

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>