yices2: Resolve improper library installation/ldconfig execution
This recipe would improperly install a symlink as the current user:

  WARNING: yices2-2.6.2+gitAUTOINC+e51be0d6ec-r0 do_package_qa: QA
  Issue: yices2: /usr/lib/libyices.so.2.6 is owned by uid 1000, which is
  the same as the user running bitbake. This may be due to host
  contamination [host-user-contaminated]

The warning is caused by the yices install target calling ldconfig,
which is not ideal for cross target use, as it would use the build hosts
ldconfig by default. Resolve this issue by disable the use of ldconfig,
and manually creating the desired symlink.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>