ghdl: Add recipe for GHDL and ghdl-yosys-plugin
This change adds support to build GHDL, however because GHDL relies on
an Ada compiler currently only native support is functional. This
requires the host to provide gnat for Ada compilation. This relies on
the gnatmake, gnatbind and gnatlink tools which must be provided by the
same compiler (and version) as gcc from the build host.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>