linux-on-litex-vexriscv: Rework to handle picolibc build
litex now uses picolibc, which is built within the litex execution.
The way this is handle is slightly odd in that picolibc depends on
libgcc headers, this means that the build needs access to a libc
toolchain. This change allows for the target toolchain to provide the
headers picolibc is depending on.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>