linux-on-litex-vexriscv: Handle litex dependency on libSDL2 for simulation
The simulation part of the litex flow allows for simulation of video
hardware via framebuffer/etc. This feature requires libSDL2 to function.
However this dependency is not cleanly conditionally configured such
that the Makefiles and code files will unconditionally attempt to
link/include the libSDL2 dependency even when no video hardware is
present in the design. For now provide the dependency for litex only
from the simulation tasks.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>