litex: Create recipes for current releases of litex and lite* components
Create release version recipes for the lite* recipes. This change splits
a number of the lite* recipes into includes with the SRCREV/PV set
within the bb. Multiple recipes are created in order to provide a
released version and a newest version based on the master branch.

By default released recipes are preferred over master version, this is
specifically enforced by the DEFAULT_PREFERENCE on the unreleased
versions.

In cases where no changes have occurred to the master branch ahead of
the released version of a recipe, drop the recipe for master changes
leaving only a release recipe.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>