nextpnr.inc: Move python3targetconfig into nextpnr.inc
Commit a682643c4ea0 added python3targetconfig to nextpnr-common to
resolve issues with nextpnr-* recipes using the wrong sysconfig and
finding the native python instead of the target python (or nativesdk
python). This however introduced the dependency into the nextpnr-chipdb
recipes which are allarch. This is problematic as these recipes are
class-target which causes python to be brought in as a DEPENDS by
python3targetconfig, a target python is not always available (e.g.
picosoc MACHINES) and would prevent the use of the native nextpnr
recipes.

Move the python3targetconfig inherit into nextpnr.inc which is only
included by the nextpnr binary target recipes.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>